Compatibility with Rainbows! 4.5.0.
Rainbows! 4.5.0 modifies the signature of the Rainbows::Response#write_headers method which now specifies that the fourth argument (body) is returned unless the connection has been hijacked via Rack hijacking functionality, in which case it should return nil.  This commit updates the Faye::WebSocket::RainbowsClient implementation to conform to this new specification.