binary frames are in exactly 1 byte-sized Integers
I recently dealt with this issue, as I was not a close enough reader of the README to realize that the description:

```
... or an `Array` of byte-sized integers (for binary frames)
```

comes to exclusively limit it to 1 byte-size integer (8 bits). I was working with 16-bit audio data and I couldn't understand why it was sounding so garbled, until a colleague and I re-read this README and understood the exclusivity of the statement. The addition of the `1` will make it clearer for future readers like me.